requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.10.6
pydantic-settings==2.7.0
anthropic==0.25.0
PyJWT==2.8.0
cryptography==42.0.0
python-dotenv==1.0.1
pathspec==0.12.1
pytest==8.1.1
pytest-asyncio==0.23.6
